+++ changelog.md
@@ -3,6 +3,7 @@
 
 - generalized IO-based functions to MonadIO. This should be non-breaking in
   most cases.
+- Added parsing functionality for JSON2 NCBI BLAST+ output
 
 0.2.0.0
 -------